requirements.txt
pandas>=2.0.0
pyodbc>=4.0.39
python-dotenv>=1.0.0   # carga variables desde .env en local
